Naming fix
